* hypernext:
  Saml orgid usercreation (#5048)
  combine reloadElement and reloadElements (#5087)
  use composer run-script for psalm and phpstan in package.json
  use nameid in SAML SLO request (#5077)
  Avoid unnecessary casting, use PDO::PARAM_X more cautiously (#5086)

+++ src/commands/ImportResources.php
@@ -52,7 +52,7 @@ class ImportResources extends Command
         $userid = (int) $input->getArgument('userid');
         $filePath = $this->Fs->getPath($input->getArgument('file'));
         $uploadedFile = new UploadedFile($filePath, 'input.eln', null, null, true);
-        $teamid = (int) (new UsersHelper($userid))->getTeamsFromUserid()[0]['id'];
+        $teamid = (new UsersHelper($userid))->getTeamsFromUserid()[0]['id'];
         $Eln = new Eln(new Users($userid, $teamid), sprintf('items:%d', $categoryId), BasePermissions::Team->toJson(), BasePermissions::User->toJson(), $uploadedFile, Storage::CACHE->getStorage()->getFs());
         $Eln->import();
